Your commit message does not fit with the provided diff file. Here is an appropriate commit based on the diff:
Commit message:
Add `.clear()` method to 'useMap' and 'useSet'

The addition of 'useMap' and the changes made to 'useSet' in source files incorporate a `.clear()` method. This allows the user to easily reset the Map or Set to its initial state without having to recreate the object. This improves the usability of these hooks and allows for more effective state management in React projects.