Update eslint-config-ethang version and simplify fullscreen check
The version of eslint-config-ethang dependency was updated from 11.5.10 to 11.5.14. This update includes several bug fixes and improvements. Corresponding updates were also made in package.json and pnpm-lock.yaml to ensure consistency.
Additionally, the way fullscreen state is initially checked in use-fullscreen.ts file was simplified for better readability and performance.